Refactored classes to use double instead of int.

+++ src/main/java/base/IShape.java
@@ -2,5 +2,5 @@ package base;
 
 public interface IShape {
 
-    int calculateArea();
+    double calculateArea();
 }
